Recomeçando. Tudo funcionando. Próxima etapa migrar models (Deus quiser vai dar certo)
