+++ core/urls.py
@@ -1,11 +1,6 @@
 from django.urls import path
-from .views import IndexView, ListarContaView, CriarContaView, ComprarDolarView, AdicionarFundosView, EditarContaView
+from .views import IndexView
 
 urlpatterns = [
     path('', IndexView.as_view(), name='index'),
-    path('listarconta/', ListarContaView.as_view(), name='listarconta'),
-    path('criarconta/', CriarContaView.as_view(), name='criarconta'),
-    path('editarconta/', EditarContaView.as_view(), name='editarconta'),
-    path('comprardolar/', ComprarDolarView.as_view(), name='comprardolar'),
-    path('adicionarfundos/', AdicionarFundosView.as_view(), name='adicionarfundos'),
 ]
